Make names and sheets optional in External
When a workbook has external references it doesn't necessarily reference any names. It's not important for the external to include an empty list in that case. So make `names` optional.

Likewise (but perhaps more rarely) a workbook _might_ reference _only_ names in an external workbook, and no sheet cells. So make `sheets` also optional.